Move data files from components to app/data
Separates static content data from UI components for better
organization. Data files are now importable from any part of
the app without coupling to component directories.
