+++ app/components/experience/index.ts
@@ -3,5 +3,4 @@ export { default as EmploymentCard } from './employment-card';
 export { default as ExperienceCard } from './experience-card';
 export { default as BackgroundCircles } from './background-circles';
 export { default as ExperienceTimeline } from './experience-timeline';
-export { experienceEntries } from './data';
-export type { ExperienceEntry } from './data';
+export { experienceEntries, type ExperienceEntry } from '@/data/experience';
