+++ app/data/case-studies.ts
